backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
bert-extractive-summarizer==0.10.1
keybert==0.8.0
nltk==3.8.1
flashtext==2.7
requests==2.31.0
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3
spacy==3.7.2
gunicorn==21.2.0